+++ web/packages/api/src/xcmBuilder.ts
@@ -1024,446 +1024,6 @@ export function buildAssetHubPNAReceivedXcm(
     })
 }
 
-export function buildTransferPolkadotToKusamaExportXCM(
-    registry: Registry,
-    transferTokenLocation: any,
-    totalFeeInNative: bigint,
-    feeOnDest: bigint,
-    sourceAssetHubParaId: number,
-    destAssetHubParaId: number,
-    transferAmount: bigint,
-    beneficiary: string,
-    topic: string
-) {
-    let withdrawAssetsOnSource: any[] = [
-        {
-            id: HERE_LOCATION,
-            fun: {
-                Fungible: totalFeeInNative,
-            },
-        },
-    ]
-    let reserveAssetsDepositedDest = [
-        {
-            id: dotLocationOnKusamaAssetHub,
-            fun: {
-                Fungible: totalFeeInNative,
-            },
-        },
-    ]
-    let withdrawAssetsDest: any[] = []
-
-    if (isNative(transferTokenLocation)) {
-        // If the asset transferred is DOT, only add the transfer amount to the asset
-        withdrawAssetsOnSource[0].fun.Fungible =
-            withdrawAssetsOnSource[0].fun.Fungible + transferAmount
-        reserveAssetsDepositedDest[0].fun.Fungible =
-            reserveAssetsDepositedDest[0].fun.Fungible + transferAmount
-    } else if (isKSMOnOtherConsensusSystem(transferTokenLocation)) {
-        // If the asset transferred is KSM, reanchor to KAH
-        withdrawAssetsDest = [
-            {
-                id: NATIVE_TOKEN_LOCATION,
-                fun: {
-                    Fungible: transferAmount,
-                },
-            },
-        ]
-    } else if (isEthereumAsset(transferTokenLocation)) {
-        // If the asset transferred is Ether or an ERC-20, the token location is already correct.
-        reserveAssetsDepositedDest.push({
-            id: transferTokenLocation,
-            fun: {
-                Fungible: transferAmount,
-            },
-        })
-    }
-    return registry.createType("XcmVersionedXcm", {
-        v4: [
-            {
-                withdrawAsset: withdrawAssetsOnSource,
-            },
-            {
-                buyExecution: {
-                    fees: {
-                        id: HERE_LOCATION,
-                        fun: {
-                            Fungible: totalFeeInNative,
-                        },
-                    },
-                    weightLimit: "Unlimited",
-                },
-            },
-            {
-                setAppendix: [
-                    {
-                        depositAsset: {
-                            assets: {
-                                wild: {
-                                    allCounted: 1,
-                                },
-                            },
-                            beneficiary: parachainLocation(sourceAssetHubParaId),
-                        },
-                    },
-                ],
-            },
-            {
-                exportMessage: {
-                    network: { Kusama: { network: null } },
-                    destination: { x1: [{ parachain: destAssetHubParaId }] },
-                    xcm: [
-                        {
-                            reserveAssetDeposited: reserveAssetsDepositedDest,
-                        },
-                        {
-                            withdrawAsset: withdrawAssetsDest,
-                        },
-                        { clearOrigin: null },
-                        {
-                            buyExecution: {
-                                fees: {
-                                    id: dotLocationOnKusamaAssetHub,
-                                    fun: {
-                                        Fungible: feeOnDest,
-                                    },
-                                },
-                                weight_limit: "Unlimited",
-                            },
-                        },
-                        {
-                            depositAsset: {
-                                assets: {
-                                    wild: {
-                                        allCounted: 2,
-                                    },
-                                },
-                                beneficiary: accountId32Location(beneficiary),
-                            },
-                        },
-                        {
-                            setTopic: topic,
-                        },
-                    ],
-                },
-            },
-
-            {
-                setTopic: topic,
-            },
-        ],
-    })
-}
-
-export function buildTransferKusamaToPolkadotExportXCM(
-    registry: Registry,
-    transferTokenLocation: any,
-    totalFeeInNative: bigint,
-    feeOnDest: bigint,
-    sourceAssetHubParaId: number,
-    destAssetHubParaId: number,
-    transferAmount: bigint,
-    beneficiary: string,
-    topic: string
-) {
-    let withdrawAssetsOnSource: any[] = [
-        {
-            id: HERE_LOCATION,
-            fun: {
-                Fungible: totalFeeInNative,
-            },
-        },
-    ]
-    let reserveAssetsDepositedDest = [
-        {
-            id: ksmLocationOnPolkadotAssetHub,
-            fun: {
-                Fungible: totalFeeInNative,
-            },
-        },
-    ]
-    let withdrawAssetsDest: any[] = []
-
-    if (isNative(transferTokenLocation)) {
-        // If the asset transferred is KSM, only add the transfer amount to the asset
-        withdrawAssetsOnSource[0].fun.Fungible =
-            withdrawAssetsOnSource[0].fun.Fungible + transferAmount
-        reserveAssetsDepositedDest[0].fun.Fungible =
-            reserveAssetsDepositedDest[0].fun.Fungible + transferAmount
-    } else if (isDOTOnOtherConsensusSystem(transferTokenLocation)) {
-        // If the asset transferred is DOT, reanchor to KAH
-        withdrawAssetsDest = [
-            {
-                id: NATIVE_TOKEN_LOCATION,
-                fun: {
-                    Fungible: transferAmount,
-                },
-            },
-        ]
-    } else if (isEthereumAsset(transferTokenLocation)) {
-        // If the asset transferred is Ether or an ERC-20, the token location is already correct.
-        withdrawAssetsDest.push({
-            id: transferTokenLocation,
-            fun: {
-                Fungible: transferAmount,
-            },
-        })
-    }
-    return registry.createType("XcmVersionedXcm", {
-        v4: [
-            {
-                withdrawAsset: withdrawAssetsDest,
-            },
-            {
-                buyExecution: {
-                    fees: {
-                        id: HERE_LOCATION,
-                        fun: {
-                            Fungible: totalFeeInNative,
-                        },
-                    },
-                    weightLimit: "Unlimited",
-                },
-            },
-            {
-                setAppendix: [
-                    {
-                        depositAsset: {
-                            assets: {
-                                wild: {
-                                    allCounted: 1,
-                                },
-                            },
-                            beneficiary: parachainLocation(sourceAssetHubParaId),
-                        },
-                    },
-                ],
-            },
-            {
-                exportMessage: {
-                    network: { Polkadot: { network: null } },
-                    destination: { x1: [{ parachain: destAssetHubParaId }] },
-                    xcm: [
-                        {
-                            reserveAssetDeposited: reserveAssetsDepositedDest,
-                        },
-                        {
-                            buyExecution: {
-                                fees: {
-                                    id: ksmLocationOnPolkadotAssetHub,
-                                    fun: {
-                                        Fungible: feeOnDest,
-                                    },
-                                },
-                                weight_limit: "Unlimited",
-                            },
-                        },
-                        {
-                            withdrawAsset: withdrawAssetsDest,
-                        },
-                        { clearOrigin: null },
-                        {
-                            depositAsset: {
-                                assets: {
-                                    wild: {
-                                        allCounted: 2,
-                                    },
-                                },
-                                beneficiary: accountId32Location(beneficiary),
-                            },
-                        },
-                        {
-                            setTopic: topic,
-                        },
-                    ],
-                },
-            },
-
-            {
-                setTopic: topic,
-            },
-        ],
-    })
-}
-
-export function buildPolkadotToKusamaDestAssetHubXCM(
-    registry: Registry,
-    totalFeeInNative: bigint,
-    assetHubParaId: number,
-    transferTokenLocation: any,
-    transferAmount: bigint,
-    beneficiary: string,
-    topic: string
-) {
-    let withdrawAssets: any[] = []
-    let reserveAssetsDeposited = [
-        {
-            id: dotLocationOnKusamaAssetHub,
-            fun: {
-                Fungible: totalFeeInNative,
-            },
-        },
-    ]
-    if (isNative(transferTokenLocation)) {
-        // If the asset transferred is DOT, only add the transfer amount to the asset
-        reserveAssetsDeposited[0].fun.Fungible =
-            reserveAssetsDeposited[0].fun.Fungible + transferAmount
-    } else if (isKSMOnOtherConsensusSystem(transferTokenLocation)) {
-        // If the asset transferred is KSM, reanchor to KAH
-        withdrawAssets = [
-            {
-                id: NATIVE_TOKEN_LOCATION,
-                fun: {
-                    Fungible: totalFeeInNative,
-                },
-            },
-        ]
-    } else if (isEthereumAsset(transferTokenLocation)) {
-        // If the asset transferred is Ether or an ERC-20, the token location is already correct.
-        reserveAssetsDeposited.push({
-            id: transferTokenLocation,
-            fun: {
-                Fungible: transferAmount,
-            },
-        })
-    }
-
-    return registry.createType("XcmVersionedXcm", {
-        v4: [
-            {
-                descendOrigin: { x1: [{ PalletInstance: 53 }] },
-            },
-            {
-                universalOrigin: polkadotNetwork,
-            },
-            {
-                descendOrigin: { x1: [{ parachain: assetHubParaId }] },
-            },
-            {
-                reserveAssetDeposited: reserveAssetsDeposited,
-            },
-            {
-                buyExecution: {
-                    fees: {
-                        id: dotLocationOnKusamaAssetHub,
-                        fun: {
-                            Fungible: totalFeeInNative,
-                        },
-                    },
-                    weightLimit: "Unlimited",
-                },
-            },
-            {
-                withdrawAsset: withdrawAssets,
-            },
-            { clearOrigin: null },
-            {
-                depositAsset: {
-                    assets: {
-                        wild: {
-                            allCounted: 2,
-                        },
-                    },
-                    beneficiary: accountId32Location(beneficiary),
-                },
-            },
-            {
-                setTopic: topic,
-            },
-        ],
-    })
-}
-
-export function buildKusamaToPolkadotDestAssetHubXCM(
-    registry: Registry,
-    totalFeeInNative: bigint,
-    assetHubParaId: number,
-    transferTokenLocation: any,
-    transferAmount: bigint,
-    beneficiary: string,
-    topic: string
-) {
-    let withdrawAssets: any[] = []
-    let reserveAssetsDeposited = [
-        {
-            id: ksmLocationOnPolkadotAssetHub,
-            fun: {
-                Fungible: totalFeeInNative,
-            },
-        },
-    ]
-    if (isNative(transferTokenLocation)) {
-        // If the asset transferred is KSM, only add the transfer amount to the asset
-        reserveAssetsDeposited[0].fun.Fungible =
-            reserveAssetsDeposited[0].fun.Fungible + transferAmount
-    } else if (isDOTOnOtherConsensusSystem(transferTokenLocation)) {
-        // If the asset transferred is DOT, reanchor to PAH
-        withdrawAssets = [
-            {
-                id: NATIVE_TOKEN_LOCATION,
-                fun: {
-                    Fungible: transferAmount,
-                },
-            },
-        ]
-    } else if (isEthereumAsset(transferTokenLocation)) {
-        // If the asset transferred is Ether or an ERC-20, the token location is already correct.
-        withdrawAssets = [
-            {
-                id: transferTokenLocation,
-                fun: {
-                    Fungible: transferAmount,
-                },
-            },
-        ]
-    }
-
-    return registry.createType("XcmVersionedXcm", {
-        v4: [
-            {
-                descendOrigin: { x1: [{ PalletInstance: 53 }] },
-            },
-            {
-                universalOrigin: kusamaNetwork,
-            },
-            {
-                descendOrigin: { x1: [{ parachain: assetHubParaId }] },
-            },
-            {
-                reserveAssetDeposited: reserveAssetsDeposited,
-            },
-            {
-                buyExecution: {
-                    fees: {
-                        id: ksmLocationOnPolkadotAssetHub,
-                        fun: {
-                            Fungible: totalFeeInNative,
-                        },
-                    },
-                    weightLimit: "Unlimited",
-                },
-            },
-            {
-                withdrawAsset: withdrawAssets,
-            },
-            { clearOrigin: null },
-            {
-                depositAsset: {
-                    assets: {
-                        wild: {
-                            allCounted: 2,
-                        },
-                    },
-                    beneficiary: accountId32Location(beneficiary),
-                },
-            },
-            {
-                setTopic: topic,
-            },
-        ],
-    })
-}
-
 export function buildExportXcmForERC20(
     registry: Registry,
     ethChainId: number,
@@ -1694,36 +1254,6 @@ export function buildExportXcmForPNA(
     })
 }
 
-export function isEthereumAsset(location: any): boolean {
-    if (location.parents !== 2 || !location.interior) return false
-
-    const interior = location.interior
-
-    const kind = Object.keys(interior).find(
-        (k) => k.toLowerCase() === "x1" || k.toLowerCase() === "x2"
-    )
-
-    if (!kind) return false
-
-    const values = interior[kind]
-    if (!Array.isArray(values) || values.length === 0) return false
-
-    const consensus = values[0]
-
-    const consensusKey = Object.keys(consensus || {}).find(
-        (k) => k.toLowerCase() === "globalconsensus"
-    )
-
-    if (!consensusKey) return false
-
-    const consensusValue = consensus[consensusKey]
-
-    return (
-        typeof consensusValue === "object" &&
-        Object.keys(consensusValue).some((k) => k.toLowerCase() === "ethereum")
-    )
-}
-
 export function isKSMOnOtherConsensusSystem(location: any) {
     return matchesConsensusSystem(location, "Kusama")
 }
@@ -1751,6 +1281,36 @@ function matchesConsensusSystem(location: any, expectedSystem: string): boolean 
     return (
         typeof consensusValue === "object" &&
         Object.keys(consensusValue).some((k) => k.toLowerCase() === expectedSystem.toLowerCase())
+    )
+}
+
+export function isEthereumAsset(location: any): boolean {
+    if (location.parents !== 2 || !location.interior) return false
+
+    const interior = location.interior
+
+    const kind = Object.keys(interior).find(
+        (k) => k.toLowerCase() === "x1" || k.toLowerCase() === "x2"
+    )
+
+    if (!kind) return false
+
+    const values = interior[kind]
+    if (!Array.isArray(values) || values.length === 0) return false
+
+    const consensus = values[0]
+
+    const consensusKey = Object.keys(consensus || {}).find(
+        (k) => k.toLowerCase() === "globalconsensus"
+    )
+
+    if (!consensusKey) return false
+
+    const consensusValue = consensus[consensusKey]
+
+    return (
+        typeof consensusValue === "object" &&
+        Object.keys(consensusValue).some((k) => k.toLowerCase() === "ethereum")
     )
 }
